Delete my account completed

+++ pages/signout.js
@@ -1,4 +1,4 @@
-import { useEffect } from "react";
+import { useEffect, useState } from "react";
 import { useToast } from "@chakra-ui/react";
 import { useRouter } from "next/router";
 
@@ -6,20 +6,33 @@ function SignOut()
 {
     const toast = useToast();
     const router = useRouter();
+    const [printed, setPrinted] = useState(false);
 
     useEffect(() => {
-        localStorage.clear();
-        
-        toast({
-            title: "Signed Out!",
-            description: "You have successfully signed out!",
-            status: "info",
-            timeout: 500,
-            isClosable: true
-        });
 
-        router.push("/login");
-    }, []);
+        if (!router.isReady) {
+            return;
+        }
+
+        if (!printed)
+        {
+            localStorage.clear();
+            toast({
+                title: "Signed Out!",
+                description: "You have successfully signed out!",
+                status: "info",
+                timeout: 2000,
+                isClosable: true
+            });
+            setPrinted(true);
+            router.push("/login");
+        }
+
+    });
+
+    return (
+        <></>
+    );
 }
 
 export default SignOut;
